ai-code-agent-adk/requirements.txt
google-adk
uvicorn
gunicorn
fastapi # get_fast_api_app likely uses FastAPI internally